refactor: restructure gel and gelv2 folders and files
