+++ lib/boorus/gelbooru/artists/artists.dart
@@ -1,0 +1,2 @@
+export 'gelbooru_artist_page.dart';
+export 'gelbooru_artist_providers.dart';
