+++ lib/boorus/gelbooru/artists/gelbooru_artist_page.dart
@@ -5,14 +5,14 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 // Project imports:
-import 'package:boorusama/boorus/gelbooru_v2/feats/posts/posts_v2.dart';
+import 'package:boorusama/boorus/gelbooru/posts/posts.dart';
 import 'package:boorusama/core/feats/boorus/boorus.dart';
 import 'package:boorusama/core/feats/tags/tags.dart';
-import 'package:boorusama/core/scaffolds/artist_page_scaffold.dart';
+import 'package:boorusama/core/scaffolds/scaffolds.dart';
 import 'package:boorusama/functional.dart';
 
-class GelbooruV2ArtistPage extends ConsumerWidget {
-  const GelbooruV2ArtistPage({
+class GelbooruArtistPage extends ConsumerWidget {
+  const GelbooruArtistPage({
     super.key,
     required this.artistName,
   });
@@ -26,7 +26,7 @@ class GelbooruV2ArtistPage extends ConsumerWidget {
     return ArtistPageScaffold(
       artistName: artistName,
       fetcher: (page, selectedCategory) =>
-          ref.read(gelbooruV2ArtistCharacterPostRepoProvider(config)).getPosts(
+          ref.read(gelbooruArtistCharacterPostRepoProvider(config)).getPosts(
                 queryFromTagFilterCategory(
                   category: selectedCategory,
                   tag: artistName,
