+++ lib/boorus/gelbooru/comments/gelbooru_comment_page.dart
@@ -1,0 +1,29 @@
+// Flutter imports:
+import 'package:flutter/material.dart';
+
+// Package imports:
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+// Project imports:
+import 'package:boorusama/core/feats/boorus/boorus.dart';
+import 'package:boorusama/core/scaffolds/scaffolds.dart';
+import 'comments.dart';
+
+class GelbooruCommentPage extends ConsumerWidget {
+  const GelbooruCommentPage({
+    super.key,
+    required this.postId,
+  });
+
+  final int postId;
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final config = ref.watchConfig;
+    return CommentPageScaffold(
+      postId: postId,
+      fetcher: (postId) =>
+          ref.watch(gelbooruCommentRepoProvider(config)).getComments(postId),
+    );
+  }
+}
