+++ lib/boorus/gelbooru/comments/gelbooru_comment_provider.dart
@@ -2,7 +2,7 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 // Project imports:
-import 'package:boorusama/boorus/gelbooru/feats/comments/comments.dart';
+import 'package:boorusama/boorus/gelbooru/comments/comments.dart';
 import 'package:boorusama/boorus/gelbooru/gelbooru.dart';
 import 'package:boorusama/core/feats/boorus/boorus.dart';
