+++ lib/boorus/gelbooru/posts/posts.dart
@@ -1,3 +1,3 @@
 export 'gelbooru_post.dart';
 export 'gelbooru_post_parser.dart';
-export 'posts_provider.dart';
+export 'gelbooru_posts_provider.dart';
