+++ lib/boorus/gelbooru_v2/artists/artists.dart
@@ -1,0 +1,2 @@
+export 'gelbooru_v2_artist_page.dart';
+export 'gelbooru_v2_artist_provider.dart';
