+++ lib/boorus/gelbooru_v2/comments/comments.dart
@@ -1,2 +1,3 @@
+export 'gelbooru_v2_comment_page.dart';
 export 'gelbooru_v2_comment_provider.dart';
 export 'gelbooru_v2_comment_repository.dart';
