+++ lib/boorus/gelbooru_v2/comments/gelbooru_v2_comment_provider.dart
@@ -2,7 +2,7 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 // Project imports:
-import 'package:boorusama/boorus/gelbooru_v2/feats/comments/comments_v2.dart';
+import 'package:boorusama/boorus/gelbooru_v2/comments/comments.dart';
 import 'package:boorusama/boorus/gelbooru_v2/gelbooru_v2.dart';
 import 'package:boorusama/core/feats/boorus/boorus.dart';
