+++ lib/boorus/gelbooru_v2/gelbooru_v2.dart
@@ -7,8 +7,9 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 // Project imports:
 import 'package:boorusama/boorus/booru_builder.dart';
 import 'package:boorusama/boorus/danbooru/danbooru.dart';
-import 'package:boorusama/boorus/gelbooru_v2/feats/comments/comments_v2.dart';
-import 'package:boorusama/boorus/gelbooru_v2/feats/posts/posts_v2.dart';
+import 'package:boorusama/boorus/gelbooru_v2/artists/artists.dart';
+import 'package:boorusama/boorus/gelbooru_v2/comments/comments.dart';
+import 'package:boorusama/boorus/gelbooru_v2/posts/posts_v2.dart';
 import 'package:boorusama/boorus/providers.dart';
 import 'package:boorusama/clients/gelbooru/gelbooru_v2_client.dart';
 import 'package:boorusama/clients/gelbooru/types/note_dto.dart';
@@ -19,15 +20,13 @@ import 'package:boorusama/core/feats/notes/notes.dart';
 import 'package:boorusama/core/feats/posts/posts.dart';
 import 'package:boorusama/core/feats/tags/tags.dart';
 import 'package:boorusama/core/scaffolds/scaffolds.dart';
-import 'package:boorusama/foundation/caching/caching.dart';
 import 'package:boorusama/foundation/networking/networking.dart';
 import 'package:boorusama/foundation/path.dart';
 import 'package:boorusama/functional.dart';
-import 'pages/create_gelbooru_v2_config_page.dart';
-import 'pages/gelbooru_v2_artist_page.dart';
-import 'pages/gelbooru_v2_home_page.dart';
-import 'pages/gelbooru_v2_post_details_desktop_page.dart';
-import 'pages/gelbooru_v2_post_details_page.dart';
+import 'create_gelbooru_v2_config_page.dart';
+import 'home/gelbooru_v2_home_page.dart';
+import 'posts/gelbooru_v2_post_details_desktop_page.dart';
+import 'posts/gelbooru_v2_post_details_page.dart';
 
 const kGelbooruV2CustomDownloadFileNameFormat =
     '{id}_{md5:maxlength=8}.{extension}';
@@ -91,35 +90,6 @@ final gelbooruV2TagsFromIdProvider =
         .toList();
   },
 );
-
-final gelbooruV2ArtistPostRepo =
-    Provider.family<PostRepository<GelbooruV2Post>, BooruConfig>((ref, config) {
-  return PostRepositoryCacher(
-    keyBuilder: (tags, page, {limit}) => '${tags.join('-')}_${page}_$limit',
-    repository: ref.watch(gelbooruV2PostRepoProvider(config)),
-    cache: LruCacher(capacity: 100),
-  );
-});
-
-final gelbooruV2ArtistPostsProvider = FutureProvider.autoDispose
-    .family<List<GelbooruV2Post>, String?>((ref, artistName) async {
-  return ref
-      .watch(gelbooruV2ArtistPostRepo(ref.watchConfig))
-      .getPostsFromTagWithBlacklist(
-        tag: artistName,
-        blacklist: ref.watch(blacklistTagsProvider(ref.watchConfig)),
-      );
-});
-
-final gelbooruV2ChildPostsProvider = FutureProvider.autoDispose
-    .family<List<GelbooruV2Post>, int>((ref, parentId) async {
-  return ref
-      .watch(gelbooruV2ArtistPostRepo(ref.watchConfig))
-      .getPostsFromTagWithBlacklist(
-        tag: 'parent:$parentId',
-        blacklist: ref.watch(blacklistTagsProvider(ref.watchConfig)),
-      );
-});
 
 final gelbooruV2NoteRepoProvider =
     Provider.family<NoteRepository, BooruConfig>((ref, config) {
@@ -337,25 +307,6 @@ class GelbooruV2SearchPage extends ConsumerWidget {
       ),
       fetcher: (page, tags) =>
           ref.watch(gelbooruV2PostRepoProvider(config)).getPosts(tags, page),
-    );
-  }
-}
-
-class GelbooruV2CommentPage extends ConsumerWidget {
-  const GelbooruV2CommentPage({
-    super.key,
-    required this.postId,
-  });
-
-  final int postId;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final config = ref.watchConfig;
-    return CommentPageScaffold(
-      postId: postId,
-      fetcher: (postId) =>
-          ref.watch(gelbooruV2CommentRepoProvider(config)).getComments(postId),
     );
   }
 }
